fix: Correct order of value in 'GameTag'
When 'mechanics' option is blank during command 'Search Card', nothing is searched.
The reason is that default of 'GameTag' is not 'INVALID'.
So fixed it by changing order of value in 'GameTag'.
Because default of enum class is first value of that in accordance with 'StrToEnum' function in 'GameEnums.hpp'.

Fixes #449

Repported-by: JinDDeok